+++ changelog.md
@@ -1,1 +1,6 @@
-- Port to 26.2
+## 26.3.0.0
+- Ported to 26.3
+- Added `resources_poplar_saplings`
+- Added `resources_poplar_leaves`
+- Added 2 new field to tree type `particle` and `ambient_leaves_block_sound_player`
+- Tree Grower can be registered with config file or API now